Add llms-full.txt file
